[run_rocm_test] - Error out if nPSDB does not have test package pre-installed

+++ bin/run_rocm_test.sh
@@ -155,6 +155,10 @@ if [ "$aomp" != 1 ]; then
   if [ "$SKIP_TEST_PACKAGE" != 1 ] && [ "$TEST_BRANCH" == "" ]; then
     git --no-pager log -1
     if [ ! -e "$ROCMINF/share/openmp-extras/tests/bin/run_rocm_test.sh" ]; then
+      if [ "$EPSDB" == "1" ]; then
+        echo "Error: nPSDB should have the openmp-extras-tests package installed before this test step."
+        exit 1
+      fi
       rm -rf $tmpdir
       mkdir -p $tmpdir
       # Determine OS and download package not using sudo.
